chore: plane boarding simulation reconfiguration
Due to the necessity of having a more consistent naming convention for
the DRL part of the project, there has been a code reordering. Based on
the concentration of all the code necessary for running the plane
boarding simulation in a single file

related: #5
